Split BLE functions out of Root class and into new Root GATT class
